windows-app: fix the DWM and registry calls, and type-check Windows from CI
CI caught two mistakes in the new cfg(windows) code, both from writing
against a newer `windows` crate API than the 0.57 this project pins: HWND is
a newtype over isize there, not a pointer, and RegOpenKeyExW takes uloptions
as a plain u32 rather than an Option.

The underlying problem is that none of that code could be checked without a
full Windows build. Installing the x86_64-pc-windows-msvc target makes
`cargo clippy --target x86_64-pc-windows-msvc` type-check it from Linux in
seconds, so add that as a step on the fast cross-platform job. It would have
caught both of these — and it immediately found a third, pre-existing:
get_blockalign() already returns u32, so the cast in the chunk-size
calculation was redundant and invisible to every check that had ever run.

+++ windows-app/src/capture/mod.rs
@@ -35,6 +35,8 @@ pub struct CapturedChunk {
 
 #[derive(Debug, thiserror::Error)]
 pub enum CaptureError {
+    // Only ever constructed by the non-Windows stub below.
+    #[cfg_attr(target_os = "windows", allow(dead_code))]
     #[error("WASAPI loopback capture is only implemented on Windows")]
     UnsupportedPlatform,
     #[cfg(target_os = "windows")]
@@ -279,8 +281,8 @@ mod windows_impl {
             // that this can change without a restart (see module doc).
             let chunk_ms = state.latency_mode().chunk_ms();
             let bytes_per_chunk =
-                (wave_format.get_blockalign() as u32 * wave_format.get_samplespersec() * chunk_ms
-                    / 1000) as usize;
+                (wave_format.get_blockalign() * wave_format.get_samplespersec() * chunk_ms / 1000)
+                    as usize;
 
             while bytes_per_chunk > 0 && sample_queue.len() >= bytes_per_chunk {
                 let pcm: Vec<u8> = sample_queue.drain(..bytes_per_chunk).collect();
